Bug 1060544 - Part 1: allow persisting/restoring home panel state r=liuche
All home panel state is currently discarded when you navigate to another page.
We want to be able to restore state (e.g. to return to the currently opened bookmarks folder,
instead of returning to the root of the bookmarks folder). This commit adds storage and a framework
to allow homepanels to persist data.

MozReview-Commit-ID: GWDX7UZrIZs